feat: coloca categorias do cardapio de maneira horizontal
